Add 'install' source script.
Copied vigorously from the aur `wl' package:

  wl/aur/broadcom-wl

Note, this install script is needed to set up proper dependencies so
that the module is auto-loaded... or something. Have a look at the
source and figure it out yourself :-).

Root yourself, then issue this command:

  $ time source ./install

Note, the ./-prefix is necessary as root, I believe. The time-prefix is
just for lulz.

Reboot and profit.