bugfix the kernel panic
For more information & the original patch, see the discussion at
https://bbs.archlinux.org/viewtopic.php?id=192287

TODO: keep track of this issue & perhaps try to apply the real patch
file (for giving credit where credit is due).

Prior to this patch, kernel would panic within a few seconds of trying
to open an SSH connection. With this patch, this initial behavior seems
to be gone. My testing is thusfar limited, but I have confidence in my
community ;-).