MARKER - support for circa linux-4.5.{something}
This marks the baseline which worked for the ARCH linux version prior to
the linux-4.6.1-2-ARCH upgrade. (Without looking into the exact version,
I believe it was circa linux-4.5.{something}).

This codebase is being abandoned in favor of more-robust / upstream
solutions.

Users of this project may look into AUR-provided solutions, as one
possible alternative.